Conflicts:
	engine/class_modules/sc_death_knight.cpp
	engine/class_modules/sc_druid.cpp
	engine/class_modules/sc_mage.cpp
	engine/class_modules/sc_monk.cpp
	engine/class_modules/sc_paladin.cpp
	engine/class_modules/sc_rogue.cpp
	engine/class_modules/sc_warlock.cpp
	engine/class_modules/sc_warrior.cpp
	engine/dbc/sc_const_data.cpp
	engine/sim/sc_scaling.cpp
	engine/simulationcraft.hpp
	profiles/PreRaid/Druid_Balance_PreRaid.simc
	profiles/PreRaid/Mage_Fire_PreRaid.simc
	profiles/Tier15H/Druid_Balance_T15H.simc
	profiles/Tier15H/Mage_Fire_T15H.simc
	profiles/Tier15H/Paladin_Protection_T15H.simc
	profiles/Tier15N/Druid_Balance_T15N.simc
	profiles/Tier15N/Mage_Fire_T15N.simc
	profiles/Tier15N/Paladin_Protection_T15N.simc
	profiles/Tier16H/Druid_Balance_T16H.simc
	profiles/Tier16H/Mage_Fire_T16H.simc
	profiles/Tier16H/Paladin_Protection_T16H.simc
	profiles/Tier16N/Druid_Balance_T16N.simc
	profiles/Tier16N/Mage_Fire_T16N.simc
	profiles/Tier16N/Paladin_Protection_T16N.simc
	vs/simcqt_vs2013.vcxproj

+++ engine/sim/sc_reforge_plot.cpp
@@ -309,8 +309,8 @@ double reforge_plot_t::progress( std::string& phase, std::string* detailed )
       phase += " to ";
   }
 
-  double total_iter = num_stat_combos * ( reforge_plot_iterations > 0 ? reforge_plot_iterations : sim -> iterations );
-  double reforge_iter = current_stat_combo * ( reforge_plot_iterations > 0 ? reforge_plot_iterations : sim -> iterations );
+  int total_iter = num_stat_combos * ( reforge_plot_iterations > 0 ? reforge_plot_iterations : sim -> iterations );
+  int reforge_iter = current_stat_combo * ( reforge_plot_iterations > 0 ? reforge_plot_iterations : sim -> iterations );
 
   if ( current_reforge_sim && current_reforge_sim -> current_iteration > 0 )
   {
@@ -325,7 +325,7 @@ double reforge_plot_t::progress( std::string& phase, std::string* detailed )
 
   sim -> detailed_progress( detailed, (int) reforge_iter, (int) total_iter );
 
-  return reforge_iter / total_iter;
+  return static_cast<double>(reforge_iter) / static_cast<double>(total_iter);
 }
 
 // reforge_plot_t::create_options ===========================================
